# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Jaga Asset Management API",
        "contact": {}
    },
    "paths": {
        "/v1/asset-categories": {
            "get": {
                "description": "Retrieve a list of asset categories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset Categories"
                ],
                "summary": "Get list of asset categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetAssetCategoriesResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new asset category with a specified name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset Categories"
                ],
                "summary": "Create a new asset category",
                "parameters": [
                    {
                        "description": "Asset category request",
                        "name": "assetCategoryRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAssetCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAssetCategoryResponse"
                        }
                    }
                }
            }
        },
        "/v1/asset-categories/{id}": {
            "get": {
                "description": "Retrieve an asset category by its unique ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset Categories"
                ],
                "summary": "Get asset category by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetAssetCategoryByIDResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing asset category by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset Categories"
                ],
                "summary": "Update an asset category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update asset category request",
                        "name": "assetCategoryRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAssetCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAssetCategoryResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an asset category by its unique ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset Categories"
                ],
                "summary": "Delete an asset category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.DeleteAssetCategoryResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AssetCategoryDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "dto.CreateAssetCategoryRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                }
            }
        },
        "dto.CreateAssetCategoryResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.DeleteAssetCategoryResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.GetAssetCategoriesResponse": {
            "type": "object",
            "properties": {
                "asset_categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.AssetCategoryDTO"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.GetAssetCategoryByIDResponse": {
            "type": "object",
            "properties": {
                "asset_category": {
                    "$ref": "#/definitions/dto.AssetCategoryDTO"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateAssetCategoryRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                }
            }
        },
        "dto.UpdateAssetCategoryResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}